feat!: extracted common Requester class from Message protobufs
Refactored the protobuf structure for message passing by introducing a new common.proto file. This includes the definition for a standalone Requester class, which was previously duplicated as an inner class for FileDownloadMessage and UserNotificationMessage in file_download.proto and notification.proto respectively. Changed the related python files to use this common Requester message. This improvement allows for less redundancy in the codebase and offers better code maintainability for any updates to the Requester definition in the future. Also updated Makefile to generate the python files for the new common.proto file.